Be more strict about unicode/bytes conversions
Previously this library relied upon duck-typed behavior, instead of
following precise schemas for the conversions between ctypes structs and
dictionaries of python objects. This led to unicode decode errors, as
bytestrings not representing human readable content were being decoded.

This fixes this issue in the following ways:

- Add `to_dict` methods to each relevant struct class that knows exactly
which fields are bytes and which are strings, and handles all
conversions appropriately.
- Make `ensure_bytes` and `ensure_string` more strict in terms of input
and output.